+++ app.yaml
@@ -17,7 +17,8 @@ handlers:
 # [START views]
 - url: /search.*
   script: Routes.app
-
+- url: /viewall
+  script: Routes.app
 - url: /viewstream
   script: Routes.app
 - url: /create
