+++ source/Main.py
@@ -75,9 +75,9 @@ class ManagePage(webapp2.RequestHandler):
             self.redirect("/")
             return
 
-        
-        
-        
+
+
+
         # get streams owned by this user
         myuser = ndb.Key('StreamUser', stream_user.key.id())
         user_streams = Stream.query(Stream.owner == myuser).fetch()
@@ -92,18 +92,18 @@ class ManagePage(webapp2.RequestHandler):
                           'id': stream.key.id()}
             owned_streams.append(streamDict)
 
-            
-            
-            
+
+
+
         # get streams subscribed by this user
         user_subscriptions = StreamSubscriber.query(StreamSubscriber.user == stream_user.key).fetch()
         print("\nstreams subscribed to by {0}: {1}".format(user.nickname(), user_subscriptions))
-        
+
         streamKeyList = []
         for sub in user_subscriptions:
             streamKeyList.append(sub.stream.get())
-        
-        subbed_streams = []        
+
+        subbed_streams = []
         for stream in streamKeyList:
             newestDate = ""
             if len(stream.items) > 0:
@@ -130,43 +130,18 @@ class ManagePage(webapp2.RequestHandler):
         self.response.write(template.render(path, template_values))
 
 
-class ViewAllStreamsPage(webapp2.RequestHandler):
-    def get(self):
-
-        user = users.get_current_user()
-
-        if user:
-            nickname = user.nickname()
-            login_url = users.create_logout_url('/')
-            login_text = 'Sign out'
-        else:
-            self.redirect("/")
-            return
-
-        template_values = {
-            'html_template': 'MasterTemplate.html',
-            'user': user,
-            'login_url': login_url,
-            'login_text': login_text,
-            'app': app_identity.get_application_id()}
-
-        self.response.content_type = 'text/html'
-        path = os.path.join(os.path.dirname(__file__), '../templates/ViewAll.html')
-        self.response.write(template.render(path, template_values))
-
-        
 class TrendingPage(webapp2.RequestHandler):
     def post(self):
         user = users.get_current_user()
         stream_user = ndb.Key('StreamUser', user.user_id()).get()
-        
+
         param_string = self.request.get('freq')
-        
+
         stream_user.update_email_freq(param_string)
 
         self.redirect("/trending")
-        
-    
+
+
     def get(self):
 
         user = users.get_current_user()
@@ -181,15 +156,15 @@ class TrendingPage(webapp2.RequestHandler):
 
         response = urllib2.urlopen('http://{0}/services/crontrends'.format(os.environ['HTTP_HOST']))
         returnValue = json.loads("".join(response.readlines()))
-        
-        
+
+
         streamList = []
         for stream in returnValue.get('trendingStreams'):
             thisStream = Stream.get_by_id(stream.get('streamKeyID'))
             streamDict = {'stream': thisStream, 'trendViews': int(stream.get('recentViews')), 'id':thisStream.key.id()}
             streamList.append(streamDict)
-        
-        
+
+
         template_values = {
             'html_template': 'MasterTemplate.html',
             'user': user,
@@ -200,13 +175,12 @@ class TrendingPage(webapp2.RequestHandler):
 
         self.response.content_type = 'text/html'
         path = os.path.join(os.path.dirname(__file__), '../templates/Trends.html')
-        self.response.write(template.render(path, template_values))        
+        self.response.write(template.render(path, template_values))
 
 
 # define the "app" that will be referenced from app.yaml
 app = webapp2.WSGIApplication([
     ('/', MainPage),
     ('/manage', ManagePage),
-    ('/view', ViewAllStreamsPage),
     ('/trending', TrendingPage)
 ], debug=True)
