+++ templates/MasterTemplate.html
@@ -27,7 +27,7 @@
         <div class="topnav">
             <a href="manage">Manage</a>
             <a href="create">Create</a>
-            <a href="view">View All</a>
+            <a href="viewall">View All</a>
             <a href="search">Search</a>
             <a href="trending">Trending</a>
             <a href="social">Social</a>
